+++ src/main/java/com/green/ca2sa/auth/AuthController.java
@@ -1,9 +1,9 @@
-package com.green.ca2sa.email;
+package com.green.ca2sa.auth;
 
 import com.green.ca2sa.common.model.ResultResponse;
-import com.green.ca2sa.email.model.AuthCodeDto;
-import com.green.ca2sa.email.model.EmailSendCodeReq;
-import com.green.ca2sa.email.model.EmailVerifyCodeReq;
+import com.green.ca2sa.auth.model.AuthCodeDto;
+import com.green.ca2sa.auth.model.SendEmailAuthCodeReq;
+import com.green.ca2sa.auth.model.VerifyEmailAuthCodeReq;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -15,15 +15,15 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("email-auth")
 @Tag(name = "이메일 인증")
-public class EmailController {
-    private final EmailService emailService;
+public class AuthController {
+    private final AuthService authService;
     private final AuthCodeService authCodeService;
 
     @PostMapping("send-code")
     @Operation(description = "이메일 인증 코드 발송")
-    public ResultResponse<Boolean> postSendCode(@RequestBody EmailSendCodeReq p) {
+    public ResultResponse<Boolean> postSendCode(@RequestBody SendEmailAuthCodeReq p) {
         AuthCodeDto authCodeDto = authCodeService.generateAuthCode(p);
-        boolean result = emailService.sendCodeToEmail(p.getEmail(), "CA2SA 이메일 인증 코드 안내", authCodeDto);
+        boolean result = authService.sendCodeToEmail(p.getEmail(), "CA2SA 이메일 인증 코드 안내", authCodeDto);
         return ResultResponse.<Boolean>builder()
                 .resultMessage(String.format("전송 %s", result ? "성공" : "실패"))
                 .resultData(result)
@@ -32,7 +32,7 @@ public class EmailController {
 
     @PostMapping("verify-code")
     @Operation(description = "이메일 인증 코드 검사")
-    public ResultResponse<Boolean> postVerifyCode(@RequestBody EmailVerifyCodeReq p) {
+    public ResultResponse<Boolean> postVerifyCode(@RequestBody VerifyEmailAuthCodeReq p) {
         boolean result = authCodeService.validateAuthCode(p);
         return ResultResponse.<Boolean>builder()
                 .resultMessage(String.format("인증 %s", result ? "성공" : "실패"))
