+++ src/main/java/com/green/ca2sa/auth/model/AuthCodeDto.java
@@ -1,4 +1,4 @@
-package com.green.ca2sa.email.model;
+package com.green.ca2sa.auth.model;
 
 import lombok.Builder;
 import lombok.Getter;
