+++ src/main/java/com/green/ca2sa/auth/model/SendEmailAuthCodeReq.java
@@ -1,4 +1,4 @@
-package com.green.ca2sa.email.model;
+package com.green.ca2sa.auth.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Getter
 @Setter
 @ToString
-public class EmailSendCodeReq {
+public class SendEmailAuthCodeReq {
     @Schema(title = "인증 받을 이메일", requiredMode = Schema.RequiredMode.REQUIRED)
     private String email;
 }
